# backend/swagger.yaml
openapi: 3.0.3
info:
  title: DataCircles CRM API
  description: |
    **DataCircles CRM + Invoicing Platform API Documentation**

    A comprehensive multi-tenant CRM and invoicing system with seat-based subscription management, OAuth authentication, and advanced document conversion capabilities.

    ## Key Features
    - Multi-tenant architecture with complete data isolation
    - OAuth 2.0 authentication (Google, Facebook, GitHub, LinkedIn) + Phone OTP
    - Seat-based subscription billing with prorated seat additions
    - Document conversion engine (Invoice ↔ Proforma ↔ Quotation ↔ Delivery Challan)
    - GST tax calculation (IGST/CGST+SGST)
    - Role-based access control (Admin/Staff)
    - Plan-based feature restrictions
    - Bulk import with CSV mapping

    ## Authentication
    All protected endpoints require authentication via:
    - **Bearer Token**: `Authorization: Bearer <jwt_token>` (for OAuth users)
    - **Phone Token**: `x-phone-token: <jwt_token>` (for phone-authenticated users)

    ## Rate Limiting
    - Standard endpoints: 100 requests per 15 minutes per IP
    - OTP endpoints: Special rate limiting to prevent abuse

    ## Pagination
    Most list endpoints support pagination with query parameters:
    - `page`: Page number (default: 1)
    - `limit`: Items per page (default: 20, max: 100)
    - `sort`: Sort field (prefix with `-` for descending)

    ## Multi-Tenancy
...
  version: 1.0.0
  contact:
    name: DataCircles Support
...
  license:
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from OAuth or phone authentication
    PhoneToken:
      type: apiKey
      in: header
      name: x-phone-token
      description: JWT token for phone-authenticated users
  schemas:
...
    Subscription:
      type: object
      properties:
        _id:
          type: string
          format: objectid
        organization:
          type: string
          format: objectid
        razorpaySubscriptionId:
          type: string
        razorpayPlanId:
          type: string
        planName:
          type: string
          enum: [starter, growth, business]
        status:
          type: string
          enum:
            [
              created,
              authenticated,
              active,
              paused,
              halted,
              cancelled,
              completed,
              expired,
            ]
        billingCycle:
          type: string
          enum: [monthly, yearly]
        pricePerUser:
          type: number
        userCount:
          type: integer
          description: Total seats purchased
        totalAmount:
          type: number
        currentPeriodStart:
          type: string
          format: date-time
        currentPeriodEnd:
          type: string
          format: date-time
        isTrialActive:
          type: boolean
        trialUsed:
          type: boolean
        isPaymentConfirmed:
          type: boolean
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
...
  responses:
    UnauthorizedError:
      description: Unauthorized - Invalid or missing authentication token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ForbiddenError:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
paths:
...
  /subscription/current:
    get:
      tags:
        - Subscriptions
      summary: Get current subscription
      description: Returns the current subscription details for the authenticated user's organization
      security:
        - BearerAuth: []
        - PhoneToken: []
      responses:
        "200":
          description: Current subscription or no subscription
          content:
            application/json:
              schema:
                type: object
                properties:
                  hasSubscription:
                    type: boolean
                  subscription:
                    $ref: "#/components/schemas/Subscription"
                  trialEligible:
                    type: boolean
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /subscription/trial:
    post:
      tags:
        - Subscriptions
      summary: Start free trial (admin only)
      description: Start a 7-day free trial on the Growth plan. Can only be used once per organization.
      security:
        - BearerAuth: []
        - PhoneToken: []
      responses:
        "200":
          description: Trial started successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  subscription:
                    $ref: "#/components/schemas/Subscription"
        "400":
          description: Trial already used
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
  /subscription/create:
    post:
      tags:
        - Subscriptions
      summary: Create new subscription (admin only)
      description: |
        Create a new subscription with Razorpay. Returns payment details for frontend integration.
        Payment must be completed for subscription to become active.
      security:
        - BearerAuth: []
        - PhoneToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - planId
                - billingCycle
                - userCount
              properties:
                planId:
                  type: string
                  enum: [starter, growth, business]
                billingCycle:
                  type: string
                  enum: [monthly, yearly]
                userCount:
                  type: integer
                  minimum: 1
                  description: Number of seats to purchase
      responses:
        "200":
          description: Subscription created, pending payment
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  subscription:
                    $ref: "#/components/schemas/Subscription"
                  paymentDetails:
                    type: object
                    description: Razorpay payment configuration for frontend
        "400":
          description: Invalid plan or organization already has subscription
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
  /subscription/update:
    put:
      tags:
        - Subscriptions
      summary: Update/upgrade subscription (admin only)
      description: |
        Update subscription plan or seat count. 
        - Upgrades require prorated payment
        - Changes are scheduled for next billing cycle by default
      security:
        - BearerAuth: []
        - PhoneToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                planId:
                  type: string
                  enum: [starter, growth, business]
                billingCycle:
                  type: string
                  enum: [monthly, yearly]
                userCount:
                  type: integer
                  minimum: 1
                immediateUpgrade:
                  type: boolean
                  description: Apply upgrade immediately with proration
      responses:
        "200":
          description: Subscription update initiated
        "400":
          description: Invalid update or no subscription found
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
  /subscription/cancel:
    post:
      tags:
        - Subscriptions
      summary: Cancel subscription (admin only)
      description: Schedule subscription cancellation at the end of current billing period
      security:
        - BearerAuth: []
        - PhoneToken: []
      responses:
        "200":
          description: Cancellation scheduled
        "400":
          description: No active subscription
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
  /subscription/verify-payment:
    post:
      tags:
        - Subscriptions
      summary: Verify subscription payment (admin only)
      description: Verify Razorpay payment signature and activate subscription
      security:
        - BearerAuth: []
        - PhoneToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - razorpay_payment_id
                - razorpay_subscription_id
                - razorpay_signature
              properties:
                razorpay_payment_id:
                  type: string
                razorpay_subscription_id:
                  type: string
                razorpay_signature:
                  type: string
      responses:
        "200":
          description: Payment verified and subscription activated
        "400":
          description: Invalid signature
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
  /subscription/payments:
    get:
      tags:
        - Subscriptions
      summary: Get payment history (admin only)
      security:
        - BearerAuth: []
        - PhoneToken: []
      responses:
        "200":
          description: List of subscription payments
          content:
            application/json:
              schema:
                type: object
                properties:
                  payments:
                    type: array
                    items:
                      type: object
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
  /subscription/webhook:
    post:
      tags:
        - Subscriptions
      summary: Razorpay webhook handler
      description: |
        Webhook endpoint for Razorpay payment events. 
        This endpoint is called automatically by Razorpay - not for manual use.
        Handles: payment.authorized, subscription.charged, subscription.cancelled, etc.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Webhook processed
        "400":
          description: Invalid webhook data

  # ==========================================
  # COMPANY ENDPOINTS
  # ==========================================
...